Publish the explanation workflow from public_workflow, not public_service
explain_schedules and build_schedules_list_diff come from
workflows/parsing/, so they belong behind public_workflow.py, next to
scheduling_extract_v2_refined_content, rather than mixed in among the
service pass-throughs. The two front-side callers follow.

public_service now takes SchedulesList from public_model for its
annotations, the way scheduling_service already does, leaving only
SCHEDULES_LIST_VERSION as an internal workflow import.
